requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
asyncpg
python-dotenv